requirements.txt
fastapi
uvicorn
pymongo
dnspython
pydantic
python-dotenv
aiohttp